+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Add new mobile design.
+### Added
+- Get switcher layout modes by schema
 
 ### Changed
 - Replace tachyons classes with design tokens.